fix: some platforms were missing the signature normalisation
Added a test for this case scenario
